Scripting Languages Examples
These are scripting language examples for games made during a class on Scripting. The python file is a map creator that walks through and attempts to solve via Right Hand Wall and Right And Run. The Perl file is a fighting game simulator that uses inputs in a rock, paper, scissor model with damage and health monitored.